+++ src/schemathesis/generation/jsonschema/__init__.py
@@ -1,4 +1,0 @@
-from schemathesis.generation.jsonschema.builder import EMPTY_STRATEGY, build
-from schemathesis.generation.jsonschema.context import Alphabet
-
-__all__ = ["EMPTY_STRATEGY", "Alphabet", "build"]