+++ src/schemathesis/specs/openapi/media_types.py
@@ -6,9 +6,6 @@ from typing import TYPE_CHECKING, Any
 from schemathesis.core import Body
 from schemathesis.core.media_types import MEDIA_TYPE_STRATEGIES
 from schemathesis.transport import SerializationContext
-from schemathesis.transport.asgi import ASGI_TRANSPORT
-from schemathesis.transport.requests import REQUESTS_TRANSPORT
-from schemathesis.transport.wsgi import WSGI_TRANSPORT
 
 if TYPE_CHECKING:
     from hypothesis import strategies as st
@@ -59,6 +56,9 @@ def register_media_type(name: str, strategy: st.SearchStrategy[bytes], *, aliase
         ```
 
     """
+    from schemathesis.transport.asgi import ASGI_TRANSPORT
+    from schemathesis.transport.requests import REQUESTS_TRANSPORT
+    from schemathesis.transport.wsgi import WSGI_TRANSPORT
 
     @REQUESTS_TRANSPORT.serializer(name, *aliases)
     @ASGI_TRANSPORT.serializer(name, *aliases)