+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.27.2...HEAD) - TBD
+## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.27.3...HEAD) - TBD
+
+## [4.27.3](https://github.com/schemathesis/schemathesis/compare/v4.27.2...v4.27.3) - 2026-09-17
 
 ### :rocket: Added